requirements.txt
# ============================================================================
# MULTIDOC-INTELLIAGENT DEPENDENCIES
# ============================================================================

# Core Dependencies
# python>=3.9

# PDF Processing
PyMuPDF>=1.23.0
camelot-py[cv]>=0.11.0
pdfplumber>=0.10.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0

# Embeddings & Vector Store
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM Integration
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# UI
streamlit>=1.28.0
streamlit-extras>=0.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Reranking
rank-bm25>=0.2.2

# Optional: GPU Support
# torch>=2.0.0
# faiss-gpu>=1.7.2

# Development Dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0
ipykernel>=6.25.0